Address review feedback
- Use baseMapItems.at(0) instead of a type cast in loadPersistedOrInitBaseMap.
- Type the loadConfig contract's config as JsonObject (rather than JsonValue),
  and validate it in defaultLoadConfig so the default implementation honestly
  conforms - a parseable-but-non-object config now raises a load error instead
  of silently no-op'ing.
- Move defaultLoadConfig (and its IS_DEVELOPMENT switch) out of Terria.ts into
  lib/Models/defaultLoadConfig.ts; re-export it from Terria so the existing
  import path keeps working.
